[MaL] moved part of NumConf and MatSpecs to separate files
