+++ MathLib/LinAlg/MatrixSpecifications.h
@@ -1,0 +1,37 @@
+/**
+ * \copyright
+ * Copyright (c) 2012-2016, OpenGeoSys Community (http://www.opengeosys.org)
+ *            Distributed under a Modified BSD License.
+ *              See accompanying file LICENSE.txt or
+ *              http://www.opengeosys.org/project/license
+ *
+ */
+
+#ifndef MATHLIB_LINALG_MATRIXSPECIFICATIONS_H
+#define MATHLIB_LINALG_MATRIXSPECIFICATIONS_H
+
+#include "GlobalMatrixVectorTypes.h"
+
+namespace MathLib
+{
+
+struct MatrixSpecifications
+{
+    MatrixSpecifications(std::size_t const nrows_, std::size_t const ncols_,
+                         std::vector<GlobalIndexType> const*const ghost_indices_,
+                         GlobalSparsityPattern const*const sparsity_pattern_)
+        : nrows(nrows_), ncols(ncols_), ghost_indices(ghost_indices_)
+        , sparsity_pattern(sparsity_pattern_)
+    {
+    }
+
+    std::size_t const nrows;
+    std::size_t const ncols;
+    std::vector<GlobalIndexType> const*const ghost_indices;
+    GlobalSparsityPattern const*const sparsity_pattern;
+};
+
+} // namespace MathLib
+
+
+#endif // MATHLIB_LINALG_MATRIXSPECIFICATIONS_H
