+++ NumLib/DOF/MatrixProviderUser.h
@@ -12,30 +12,10 @@
 
 #include <cstddef>
 
-#include "NumLib/NumericsConfig.h"
-
-namespace NumLib { class LocalToGlobalIndexMap; }
-namespace MeshLib { class Mesh; }
+#include "MathLib/LinAlg/MatrixSpecifications.h"
 
 namespace MathLib
 {
-
-struct MatrixSpecifications
-{
-    MatrixSpecifications(std::size_t const nrows_, std::size_t const ncols_,
-                         GlobalSparsityPattern const*const sparsity_pattern_,
-                         NumLib::LocalToGlobalIndexMap const*const dof_table_,
-                         MeshLib::Mesh const*const mesh_)
-        : nrows(nrows_), ncols(ncols_), sparsity_pattern(sparsity_pattern_)
-        , dof_table(dof_table_), mesh(mesh_)
-    {}
-
-    std::size_t const nrows;
-    std::size_t const ncols;
-    GlobalSparsityPattern const*const sparsity_pattern;
-    NumLib::LocalToGlobalIndexMap const*const dof_table;
-    MeshLib::Mesh const*const mesh;
-};
 
 class MatrixSpecificationsProvider
 {
